src/02: Optimize away the allocator
Yeah, the code is a lot smaller, and arguably harder to read/understand,
but we no longer have to explicitly allocate anything and worry about
freeing allocations at the end of it.

Surprisingly, the speed-up isn't that much!

The original version (which used allocators) ran in 0.029s. This new
version (without allocators) runs in 0.027s.